Bridges STEM Video Showcase Preparation
The VFX file has FX and Bridges usage recordings which can be used at different stages of the video or in a B-roll to enhance the video.

Storyboard has a basic order of events and topics that will be covered during the duration of the video. Can be filled in and updated through the development process.